+++ R/shapviz-package.R
@@ -4,7 +4,7 @@
 # Suppress R CMD check note
 #' @importFrom xgboost xgb.train
 
-globalVariables(c("from", "i", "id", "label", "to", "x", "shap",
+globalVariables(c("from", "i", "id", "label", "to", "x", "shap", "SHAP",
                   "feature", "value", "color", "Var2", "Var3", "S"))
 
 .onLoad <- function(libname, pkgname) {